got api server hooked up to front end properly

+++ server/connectDB.js
@@ -1,14 +1,19 @@
 const { mongoose } = require('mongoose')
+const dotenv = require("dotenv")
+dotenv.config();
+// console.log(process.env)
 
-const connectDB = async () => {
-    try {
-        const x = await mongoose.connect('mongodb+srv://User1:C1Y1FCRG3Bja5wX8@clusterdouglas.vgdrqt1.mongodb.net/?retryWrites=true&w=majority')
-        console.log('Connected to db');
-        mongoose.connection.db.dropDatabase();
-        console.log("Dropped db");
-    } catch (error) {
-        console.log('db error');
-    }
+const connectDB = async (input) => {
+  try {
+    const x = await mongoose.connect(process.env.DB_STRING)
+    console.log("Connected to db");
+    if (input.drop === true)
+      mongoose.connection.db.dropDatabase();
+    // console.log("Dropped db");
+    // get the data from Github 
+  } catch (error) {
+    console.log('db error');
+  }
 }
 
 module.exports = { connectDB }
